requirements.txt
fastapi
unicorn
gspread
oauth2client
python-dotenv
pydantic